requirements.txt
# Core Django
Django==4.2.26  # Security update: fixes 21 CVEs including SQL injection, DoS, log injection
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.8.2

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Celery
celery==5.3.6
celery[redis]==5.3.6
django-celery-beat==2.5.0

# Development Tools
black==25.1.0  # Security update: fixes ReDoS vulnerability
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7
pre-commit==3.6.2
watchdog==4.0.0  # For translation hot reload in development

# Testing
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pyotp==2.9.0

# Logging
python-json-logger==2.0.7

# Authentication & Authorization
django-allauth==0.61.1
argon2-cffi==23.1.0
djangorestframework==3.15.2  # Security update: fixes XSS vulnerability
djangorestframework-simplejwt==5.5.1  # Security update: fixes info disclosure
django-otp==1.3.0
qrcode==7.4.2
python-barcode==0.15.1
Pillow==11.0.0  # Security update: fixes buffer overflow vulnerability
django-guardian==2.4.0

# PDF Generation & Reporting
reportlab==4.0.9
openpyxl==3.1.2
django-import-export==3.3.7

# Accounting
django-ledger==0.6.1

# State Machine
django-fsm==2.8.1

# HTTP Requests
requests==2.32.4  # Security update: fixes cert verification bypass and .netrc credential leak

# Email
django-anymail==10.2

# SMS
twilio==8.10.3

# System Monitoring
psutil==5.9.8
django-prometheus==2.3.1
django-silk==5.0.4

# OpenTelemetry - Distributed Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0

# Admin Impersonation
django-hijack==3.4.5

# Feature Flags
django-waffle==4.1.0

# Payment Gateway
# Using stripe 4.x for compatibility with dj-stripe
stripe==4.2.0

# Cloud Storage (for backups)
boto3==1.34.51

# Encryption
cryptography==44.0.1  # Security update: fixes OpenSSL vulnerabilities
responses==0.25.8

# Translation Management
django-rosetta==0.10.0

# Persian Calendar & Date Formatting
jdatetime==5.0.0

# Form Widgets
django-widget-tweaks==1.5.0

# Bank statement parsing
ofxparse==0.21

# Asset Compression & Optimization
django-compressor==4.4
csscompressor==0.9.5
jsmin==3.0.1

# API Rate Limiting
django-ratelimit==4.1.0

# Performance Testing
locust==2.20.0

# Error Tracking & Monitoring
sentry-sdk[django]==2.20.0  # Security update: fixes env variable exposure

# Security Scanning
bandit==1.7.7
pip-audit==2.7.3

# Production WSGI Server
gunicorn==23.0.0  # Security update: fixes HTTP Request Smuggling vulnerabilities
